requirements.txt
torch>=2.0.0
numpy>=1.22.0
Pillow>=9.0.0
faiss-cpu>=1.7.0
open_clip_torch>=2.20.0
requests>=2.28.0
